+++ lib/src/v2/document.dart
@@ -1,13 +1,13 @@
-import 'package:conduit_codable/conduit_codable.dart';
-import 'package:conduit_codable/cast.dart' as cast;
-import 'package:conduit_open_api/src/object.dart';
-import 'package:conduit_open_api/src/util/list_helper.dart';
-import 'package:conduit_open_api/src/v2/metadata.dart';
-import 'package:conduit_open_api/src/v2/parameter.dart';
-import 'package:conduit_open_api/src/v2/path.dart';
-import 'package:conduit_open_api/src/v2/response.dart';
-import 'package:conduit_open_api/src/v2/schema.dart';
-import 'package:conduit_open_api/src/v2/security.dart';
+import 'package:conduit_codable_fork/conduit_codable.dart';
+import 'package:conduit_codable_fork/cast.dart' as cast;
+import 'package:conduit_open_api_fork/src/object.dart';
+import 'package:conduit_open_api_fork/src/util/list_helper.dart';
+import 'package:conduit_open_api_fork/src/v2/metadata.dart';
+import 'package:conduit_open_api_fork/src/v2/parameter.dart';
+import 'package:conduit_open_api_fork/src/v2/path.dart';
+import 'package:conduit_open_api_fork/src/v2/response.dart';
+import 'package:conduit_open_api_fork/src/v2/schema.dart';
+import 'package:conduit_open_api_fork/src/v2/security.dart';
 
 /// Represents an OpenAPI 2.0 specification.
 class APIDocument extends APIObject {
@@ -45,8 +45,7 @@ class APIDocument extends APIObject {
         "schemes": const cast.List(cast.string),
         "consumes": const cast.List(cast.string),
         "produces": const cast.List(cast.string),
-        "security":
-            const cast.List(cast.Map(cast.string, cast.List(cast.string)))
+        "security": const cast.List(cast.Map(cast.string, cast.List(cast.string)))
       };
 
   @override
@@ -68,10 +67,11 @@ class APIDocument extends APIObject {
     paths = object.decodeObjectMap("paths", () => APIPath());
     responses = object.decodeObjectMap("responses", () => APIResponse());
     parameters = object.decodeObjectMap("parameters", () => APIParameter());
-    definitions =
-        object.decodeObjectMap("definitions", () => APISchemaObject());
+    definitions = object.decodeObjectMap("definitions", () => APISchemaObject());
     securityDefinitions = object.decodeObjectMap(
-        "securityDefinitions", () => APISecurityScheme());
+      "securityDefinitions",
+      () => APISecurityScheme(),
+    );
   }
 
   @override