+++ lib/src/v2/operation.dart
@@ -1,8 +1,8 @@
-import 'package:conduit_codable/conduit_codable.dart';
-import 'package:conduit_codable/cast.dart' as cast;
-import 'package:conduit_open_api/src/object.dart';
-import 'package:conduit_open_api/src/v2/parameter.dart';
-import 'package:conduit_open_api/src/v2/response.dart';
+import 'package:conduit_codable_fork/cast.dart' as cast;
+import 'package:conduit_codable_fork/conduit_codable.dart';
+import 'package:conduit_open_api_fork/src/object.dart';
+import 'package:conduit_open_api_fork/src/v2/parameter.dart';
+import 'package:conduit_open_api_fork/src/v2/response.dart';
 
 /// Represents a HTTP operation (a path/method pair) in the OpenAPI specification.
 class APIOperation extends APIObject {
@@ -14,8 +14,7 @@ class APIOperation extends APIObject {
         "consumes": const cast.List(cast.string),
         "produces": const cast.List(cast.string),
         "schemes": const cast.List(cast.string),
-        "security":
-            const cast.List(cast.Map(cast.string, cast.List(cast.string))),
+        "security": const cast.List(cast.Map(cast.string, cast.List(cast.string))),
       };
 
   String? summary = "";