+++ lib/src/v2/response.dart
@@ -1,7 +1,7 @@
-import 'package:conduit_codable/conduit_codable.dart';
-import 'package:conduit_open_api/src/object.dart';
-import 'package:conduit_open_api/src/v2/header.dart';
-import 'package:conduit_open_api/src/v2/schema.dart';
+import 'package:conduit_codable_fork/conduit_codable.dart';
+import 'package:conduit_open_api_fork/src/object.dart';
+import 'package:conduit_open_api_fork/src/v2/header.dart';
+import 'package:conduit_open_api_fork/src/v2/schema.dart';
 
 /// Represents an HTTP response in the OpenAPI specification.
 class APIResponse extends APIObject {