+++ lib/src/v3/callback.dart
@@ -1,6 +1,6 @@
-import 'package:conduit_codable/conduit_codable.dart';
-import 'package:conduit_open_api/src/object.dart';
-import 'package:conduit_open_api/src/v3/path.dart';
+import 'package:conduit_codable_fork/conduit_codable.dart';
+import 'package:conduit_open_api_fork/src/object.dart';
+import 'package:conduit_open_api_fork/src/v3/path.dart';
 
 /// A map of possible out-of band callbacks related to the parent operation.
 ///
@@ -22,7 +22,8 @@ class APICallback extends APIObject {
     object.forEach((key, dynamic value) {
       if (value is! KeyedArchive) {
         throw ArgumentError(
-            "Invalid specification. Callback contains non-object value.");
+          "Invalid specification. Callback contains non-object value.",
+        );
       }
       paths![key] = value.decodeObject(key, () => APIPath())!;
     });