+++ lib/src/v3/operation.dart
@@ -1,14 +1,14 @@
-import 'package:conduit_codable/cast.dart' as cast;
-import 'package:conduit_codable/conduit_codable.dart';
-import 'package:conduit_open_api/src/object.dart';
-import 'package:conduit_open_api/src/v3/callback.dart';
-import 'package:conduit_open_api/src/v3/parameter.dart';
-import 'package:conduit_open_api/src/v3/request_body.dart';
-import 'package:conduit_open_api/src/v3/response.dart';
-import 'package:conduit_open_api/src/v3/security.dart';
-import 'package:conduit_open_api/src/v3/path.dart';
-import 'package:conduit_open_api/src/v3/document.dart';
-import 'package:conduit_open_api/src/v3/server.dart';
+import 'package:conduit_codable_fork/cast.dart' as cast;
+import 'package:conduit_codable_fork/conduit_codable.dart';
+import 'package:conduit_open_api_fork/src/object.dart';
+import 'package:conduit_open_api_fork/src/v3/callback.dart';
+import 'package:conduit_open_api_fork/src/v3/parameter.dart';
+import 'package:conduit_open_api_fork/src/v3/request_body.dart';
+import 'package:conduit_open_api_fork/src/v3/response.dart';
+import 'package:conduit_open_api_fork/src/v3/security.dart';
+import 'package:conduit_open_api_fork/src/v3/path.dart';
+import 'package:conduit_open_api_fork/src/v3/document.dart';
+import 'package:conduit_open_api_fork/src/v3/server.dart';
 
 /// Describes a single API operation on a path.
 class APIOperation extends APIObject {
@@ -140,15 +140,12 @@ class APIOperation extends APIObject {
     description = object.decode("description");
     id = object.decode("operationId");
     parameters = object.decodeObjects("parameters", () => APIParameter.empty());
-    requestBody =
-        object.decodeObject("requestBody", () => APIRequestBody.empty());
+    requestBody = object.decodeObject("requestBody", () => APIRequestBody.empty());
     responses = object.decodeObjectMap("responses", () => APIResponse.empty());
     callbacks = object.decodeObjectMap("callbacks", () => APICallback());
     deprecated = object.decode("deprecated");
-    security =
-        object.decodeObjects("security", () => APISecurityRequirement.empty());
-    servers =
-        object.decodeObjects("servers", () => APIServerDescription.empty());
+    security = object.decodeObjects("security", () => APISecurityRequirement.empty());
+    servers = object.decodeObjects("servers", () => APIServerDescription.empty());
   }
 
   @override
@@ -157,7 +154,8 @@ class APIOperation extends APIObject {
 
     if (responses == null) {
       throw ArgumentError(
-          "Invalid specification. APIOperation must have non-null values for: 'responses'.");
+        "Invalid specification. APIOperation must have non-null values for: 'responses'.",
+      );
     }
 
     object.encode("tags", tags);