+++ lib/src/v3/request_body.dart
@@ -1,7 +1,7 @@
-import 'package:conduit_codable/conduit_codable.dart';
-import 'package:conduit_open_api/src/object.dart';
-import 'package:conduit_open_api/src/v3/media_type.dart';
-import 'package:conduit_open_api/src/v3/schema.dart';
+import 'package:conduit_codable_fork/conduit_codable.dart';
+import 'package:conduit_open_api_fork/src/object.dart';
+import 'package:conduit_open_api_fork/src/v3/media_type.dart';
+import 'package:conduit_open_api_fork/src/v3/schema.dart';
 
 /// Describes a single request body.
 class APIRequestBody extends APIObject {
@@ -49,7 +49,8 @@ class APIRequestBody extends APIObject {
 
     if (content == null) {
       throw ArgumentError(
-          "APIRequestBody must have non-null values for: 'content'.");
+        "APIRequestBody must have non-null values for: 'content'.",
+      );
     }
 
     object.encode("description", description);