+++ lib/src/v3/security.dart
@@ -1,9 +1,9 @@
-import 'package:conduit_codable/conduit_codable.dart';
-import 'package:conduit_open_api/src/object.dart';
-import 'package:conduit_open_api/src/v3/components.dart';
-import 'package:conduit_open_api/src/v3/document.dart';
-import 'package:conduit_open_api/src/v3/operation.dart';
-import 'package:conduit_open_api/src/v3/parameter.dart';
+import 'package:conduit_codable_fork/conduit_codable.dart';
+import 'package:conduit_open_api_fork/src/object.dart';
+import 'package:conduit_open_api_fork/src/v3/components.dart';
+import 'package:conduit_open_api_fork/src/v3/document.dart';
+import 'package:conduit_open_api_fork/src/v3/operation.dart';
+import 'package:conduit_open_api_fork/src/v3/parameter.dart';
 
 enum APISecuritySchemeType { apiKey, http, oauth2, openID }
 
@@ -48,13 +48,11 @@ class APISecurityScheme extends APIObject {
 
   APISecurityScheme.http(this.scheme) : type = APISecuritySchemeType.http;
 
-  APISecurityScheme.apiKey(this.name, this.location)
-      : type = APISecuritySchemeType.apiKey;
+  APISecurityScheme.apiKey(this.name, this.location) : type = APISecuritySchemeType.apiKey;
 
   APISecurityScheme.oauth2(this.flows) : type = APISecuritySchemeType.oauth2;
 
-  APISecurityScheme.openID(this.connectURL)
-      : type = APISecuritySchemeType.openID;
+  APISecurityScheme.openID(this.connectURL) : type = APISecuritySchemeType.openID;
 
   /// The type of the security scheme.
   ///
@@ -120,7 +118,9 @@ class APISecurityScheme extends APIObject {
       case APISecuritySchemeType.oauth2:
         {
           flows = object.decodeObjectMap(
-              "flows", () => APISecuritySchemeOAuth2Flow.empty());
+            "flows",
+            () => APISecuritySchemeOAuth2Flow.empty(),
+          );
         }
         break;
       case APISecuritySchemeType.http:
@@ -136,7 +136,8 @@ class APISecurityScheme extends APIObject {
         break;
       default:
         throw ArgumentError(
-            "APISecurityScheme must have non-null values for: 'type'.");
+          "APISecurityScheme must have non-null values for: 'type'.",
+        );
     }
   }
 
@@ -146,7 +147,8 @@ class APISecurityScheme extends APIObject {
 
     if (type == null) {
       throw ArgumentError(
-          "APISecurityScheme must have non-null values for: 'type'.");
+        "APISecurityScheme must have non-null values for: 'type'.",
+      );
     }
 
     object.encode("type", APISecuritySchemeTypeCodec.encode(type));
@@ -157,7 +159,8 @@ class APISecurityScheme extends APIObject {
         {
           if (name == null || location == null) {
             throw ArgumentError(
-                "APISecurityScheme with 'apiKey' type must have non-null values for: 'name', 'location'.");
+              "APISecurityScheme with 'apiKey' type must have non-null values for: 'name', 'location'.",
+            );
           }
 
           object.encode("name", name);
@@ -168,7 +171,8 @@ class APISecurityScheme extends APIObject {
         {
           if (flows == null) {
             throw ArgumentError(
-                "APISecurityScheme with 'oauth2' type must have non-null values for: 'flows'.");
+              "APISecurityScheme with 'oauth2' type must have non-null values for: 'flows'.",
+            );
           }
 
           object.encodeObjectMap("flows", flows);
@@ -178,7 +182,8 @@ class APISecurityScheme extends APIObject {
         {
           if (scheme == null) {
             throw ArgumentError(
-                "APISecurityScheme with 'http' type must have non-null values for: 'scheme'.");
+              "APISecurityScheme with 'http' type must have non-null values for: 'scheme'.",
+            );
           }
 
           object.encode("scheme", scheme);
@@ -189,14 +194,16 @@ class APISecurityScheme extends APIObject {
         {
           if (connectURL == null) {
             throw ArgumentError(
-                "APISecurityScheme with 'openID' type must have non-null values for: 'connectURL'.");
+              "APISecurityScheme with 'openID' type must have non-null values for: 'connectURL'.",
+            );
           }
           object.encode("openIdConnectUrl", connectURL);
         }
         break;
       default:
         throw ArgumentError(
-            "APISecurityScheme must have non-null values for: 'type'.");
+          "APISecurityScheme must have non-null values for: 'type'.",
+        );
     }
   }
 }
@@ -206,12 +213,9 @@ class APISecuritySchemeOAuth2Flow extends APIObject {
   APISecuritySchemeOAuth2Flow.empty();
   APISecuritySchemeOAuth2Flow.code(
       this.authorizationURL, this.tokenURL, this.refreshURL, this.scopes);
-  APISecuritySchemeOAuth2Flow.implicit(
-      this.authorizationURL, this.refreshURL, this.scopes);
-  APISecuritySchemeOAuth2Flow.password(
-      this.tokenURL, this.refreshURL, this.scopes);
-  APISecuritySchemeOAuth2Flow.client(
-      this.tokenURL, this.refreshURL, this.scopes);
+  APISecuritySchemeOAuth2Flow.implicit(this.authorizationURL, this.refreshURL, this.scopes);
+  APISecuritySchemeOAuth2Flow.password(this.tokenURL, this.refreshURL, this.scopes);
+  APISecuritySchemeOAuth2Flow.client(this.tokenURL, this.refreshURL, this.scopes);
 
   /// The authorization URL to be used for this flow.
   ///