+++ test/v3_test.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:conduit_open_api/v3.dart';
+import 'package:conduit_open_api_fork/v3.dart';
 import 'package:dcli/dcli.dart';
 import 'package:test/test.dart';
 
@@ -11,15 +11,14 @@ void main() {
       final components = APIComponents();
       components.schemas["foo"] = APISchemaObject.string();
 
-      final ref = APISchemaObject()
-        ..referenceURI = Uri.parse("/components/schemas/foo");
+      final ref = APISchemaObject()..referenceURI = Uri.parse("/components/schemas/foo");
       final orig = components.resolve(ref);
       expect(orig, isNotNull);
       expect(orig!.type, APIType.string);
       expect(ref.type, isNull);
 
-      final APISchemaObject? constructed = components
-          .resolveUri(Uri(path: "/components/schemas/foo")) as APISchemaObject?;
+      final APISchemaObject? constructed =
+          components.resolveUri(Uri(path: "/components/schemas/foo")) as APISchemaObject?;
       expect(constructed, isNotNull);
       expect(constructed!.type, APIType.string);
     });
@@ -27,8 +26,7 @@ void main() {
     test("Invalid ref uri format throws error", () {
       final components = APIComponents();
       try {
-        components.resolve(APISchemaObject()
-          ..referenceURI = Uri.parse("#/components/schemas/foo"));
+        components.resolve(APISchemaObject()..referenceURI = Uri.parse("#/components/schemas/foo"));
         expect(true, false);
         // ignore: avoid_catching_errors
       } on ArgumentError catch (e) {
@@ -36,8 +34,7 @@ void main() {
       }
 
       try {
-        components.resolve(APISchemaObject()
-          ..referenceURI = Uri.parse("#/components/schemas"));
+        components.resolve(APISchemaObject()..referenceURI = Uri.parse("#/components/schemas"));
         expect(true, false);
         // ignore: avoid_catching_errors
       } on ArgumentError catch (e) {
@@ -45,8 +42,7 @@ void main() {
       }
 
       try {
-        components.resolve(APISchemaObject()
-          ..referenceURI = Uri.parse("/components/foobar/foo"));
+        components.resolve(APISchemaObject()..referenceURI = Uri.parse("/components/foobar/foo"));
         expect(true, false);
         // ignore: avoid_catching_errors
       } on ArgumentError catch (e) {
@@ -57,8 +53,8 @@ void main() {
     test("Nonexisting component returns null", () {
       final components = APIComponents();
       expect(
-          components.resolve(APISchemaObject()
-            ..referenceURI = Uri.parse("/components/schemas/foo")),
+          components
+              .resolve(APISchemaObject()..referenceURI = Uri.parse("/components/schemas/foo")),
           isNull);
     });
 
@@ -77,17 +73,15 @@ void main() {
         }
       });
 
-      expect(doc.components!.schemas["container"]!.referenceURI!.path,
-          "/components/schemas/string");
+      expect(
+          doc.components!.schemas["container"]!.referenceURI!.path, "/components/schemas/string");
 
       doc.components!.schemas["other"] = APISchemaObject()
         ..referenceURI = Uri(path: "/components/schemas/container");
 
       final out = doc.asMap();
-      expect(out["components"]["schemas"]["container"][r"$ref"],
-          "#/components/schemas/string");
-      expect(out["components"]["schemas"]["other"][r"$ref"],
-          "#/components/schemas/container");
+      expect(out["components"]["schemas"]["container"][r"$ref"], "#/components/schemas/string");
+      expect(out["components"]["schemas"]["other"][r"$ref"], "#/components/schemas/container");
     });
   });
 
@@ -119,8 +113,7 @@ void main() {
       expect(doc!.info.version, isNotNull);
       expect(doc!.info.description,
           "The Stripe REST API. Please see https://stripe.com/docs/api for more details.");
-      expect(doc!.info.termsOfServiceURL.toString(),
-          "https://stripe.com/us/terms/");
+      expect(doc!.info.termsOfServiceURL.toString(), "https://stripe.com/us/terms/");
       expect(doc!.info.contact!.email, "dev-platform@stripe.com");
       expect(doc!.info.contact!.name, "Stripe Dev Platform Team");
       expect(doc!.info.contact!.url.toString(), "https://stripe.com");
@@ -174,21 +167,15 @@ void main() {
 
         expect(getResponses!.length, 2);
         expect(getResponses["200"]!.content!.length, 1);
-        expect(
-            getResponses["200"]!
-                .content!["application/json"]!
-                .schema!
-                .referenceURI,
-            Uri.parse(
-                Uri.parse("#/components/schemas/transfer_reversal").fragment));
+        expect(getResponses["200"]!.content!["application/json"]!.schema!.referenceURI,
+            Uri.parse(Uri.parse("#/components/schemas/transfer_reversal").fragment));
 
         final resolvedElement = getResponses["200"]!
             .content!["application/json"]!
             .schema!
             .properties!["balance_transaction"]!
             .anyOf;
-        expect(resolvedElement!.last!.properties!["amount"]!.type,
-            APIType.integer);
+        expect(resolvedElement!.last!.properties!["amount"]!.type, APIType.integer);
       });
     });
 
@@ -216,12 +203,8 @@ void main() {
       expect(doc.components!.schemas["freeform"]!.additionalPropertyPolicy,
           APISchemaAdditionalPropertyPolicy.freeForm);
 
-      expect(
-          doc.asMap()["components"]["schemas"]["freeform"]["type"], "object");
-      expect(
-          doc.asMap()["components"]["schemas"]["freeform"]
-              ["additionalProperties"],
-          true);
+      expect(doc.asMap()["components"]["schemas"]["freeform"]["type"], "object");
+      expect(doc.asMap()["components"]["schemas"]["freeform"]["additionalProperties"], true);
     });
 
     test("Can read/emit schema object with additionalProperties={}", () {
@@ -231,21 +214,14 @@ void main() {
         "paths": <String, dynamic>{},
         "components": {
           "schemas": {
-            "freeform": {
-              "type": "object",
-              "additionalProperties": <String, dynamic>{}
-            }
+            "freeform": {"type": "object", "additionalProperties": <String, dynamic>{}}
           }
         }
       });
       expect(doc.components!.schemas["freeform"]!.additionalPropertyPolicy,
           APISchemaAdditionalPropertyPolicy.freeForm);
-      expect(
-          doc.asMap()["components"]["schemas"]["freeform"]["type"], "object");
-      expect(
-          doc.asMap()["components"]["schemas"]["freeform"]
-              ["additionalProperties"],
-          true);
+      expect(doc.asMap()["components"]["schemas"]["freeform"]["type"], "object");
+      expect(doc.asMap()["components"]["schemas"]["freeform"]["additionalProperties"], true);
     });
   });
 
@@ -256,19 +232,16 @@ void main() {
       final resp = APIResponse("Response");
 
       // when null
-      resp.addHeader(
-          "x", APIHeader(schema: APISchemaObject.string(format: "initial")));
+      resp.addHeader("x", APIHeader(schema: APISchemaObject.string(format: "initial")));
       expect(resp.headers!["x"]!.schema!.format, "initial");
 
       // add more than one
-      resp.addHeader(
-          "y", APIHeader(schema: APISchemaObject.string(format: "second")));
+      resp.addHeader("y", APIHeader(schema: APISchemaObject.string(format: "second")));
       expect(resp.headers!["x"]!.schema!.format, "initial");
       expect(resp.headers!["y"]!.schema!.format, "second");
 
       // cannot replace
-      resp.addHeader(
-          "y", APIHeader(schema: APISchemaObject.string(format: "third")));
+      resp.addHeader("y", APIHeader(schema: APISchemaObject.string(format: "third")));
       expect(resp.headers!["x"]!.schema!.format, "initial");
       expect(resp.headers!["y"]!.schema!.format, "second");
     });
@@ -297,34 +270,23 @@ void main() {
       final op = APIOperation("op", null);
 
       // when null
-      op.addResponse(200,
-          APIResponse.schema("OK", APISchemaObject.string(format: "initial")));
-      expect(op.responses!["200"]!.content!["application/json"]!.schema!.format,
-          "initial");
+      op.addResponse(200, APIResponse.schema("OK", APISchemaObject.string(format: "initial")));
+      expect(op.responses!["200"]!.content!["application/json"]!.schema!.format, "initial");
 
       // add more than one
       op.addResponse(
           400,
-          APIResponse.schema(
-              "KINDABAD", APISchemaObject.string(format: "second"), headers: {
-            "initial":
-                APIHeader(schema: APISchemaObject.string(format: "initial"))
-          }));
-      expect(op.responses!["200"]!.content!["application/json"]!.schema!.format,
-          "initial");
-      expect(op.responses!["400"]!.content!["application/json"]!.schema!.format,
-          "second");
+          APIResponse.schema("KINDABAD", APISchemaObject.string(format: "second"),
+              headers: {"initial": APIHeader(schema: APISchemaObject.string(format: "initial"))}));
+      expect(op.responses!["200"]!.content!["application/json"]!.schema!.format, "initial");
+      expect(op.responses!["400"]!.content!["application/json"]!.schema!.format, "second");
 
       // join responses when key exists
       op.addResponse(
           400,
           APIResponse.schema("REALBAD", APISchemaObject.string(format: "third"),
-              headers: {
-                "second":
-                    APIHeader(schema: APISchemaObject.string(format: "initial"))
-              }));
-      expect(op.responses!["200"]!.content!["application/json"]!.schema!.format,
-          "initial");
+              headers: {"second": APIHeader(schema: APISchemaObject.string(format: "initial"))}));
+      expect(op.responses!["200"]!.content!["application/json"]!.schema!.format, "initial");
 
       final r400 = op.responses!["400"]!;
       expect(r400.description, contains("KINDABAD"));
@@ -337,21 +299,11 @@ void main() {
     test("'addResponse' guards against null value", () {
       final op = APIOperation("op", null);
 
-      op.addResponse(
-          400,
-          APIResponse.schema(
-              "KINDABAD", APISchemaObject.string(format: "second")));
-      expect(op.responses!["400"]!.content!["application/json"]!.schema!.format,
-          "second");
+      op.addResponse(400, APIResponse.schema("KINDABAD", APISchemaObject.string(format: "second")));
+      expect(op.responses!["400"]!.content!["application/json"]!.schema!.format, "second");
 
-      op.addResponse(
-          400,
-          APIResponse.schema(
-              "REALBAD", APISchemaObject.string(format: "third")));
-      expect(
-          op.responses!["400"]!.content!["application/json"]!.schema!.oneOf!
-              .length,
-          2);
+      op.addResponse(400, APIResponse.schema("REALBAD", APISchemaObject.string(format: "third")));
+      expect(op.responses!["400"]!.content!["application/json"]!.schema!.oneOf!.length, 2);
     });
   });
 }
@@ -368,8 +320,7 @@ String fetchStripExample() {
     }
 
     fetch(
-        url:
-            'https://raw.githubusercontent.com/stripe/openapi/master/openapi/spec3.json',
+        url: 'https://raw.githubusercontent.com/stripe/openapi/master/openapi/spec3.json',
         saveToPath: config);
   }
   return config;